Basic story builder app skeleton
The goal of this patch is to provide a basic app structure
that can be used as a starting point for the story builder.

It includes a call to action (create a story) in the article
header section. It has a store and routing and a few examples
of vue files.

Change-Id: I2fe6fd0ad210fdb9b52dcaa38f3946be02b8645b