Keep search results
In the StoryBuilder, when the user goes back to the search
page to add more frames, the search is performed again even
though the query is still the same and the search results
are already present.

If the query is the same and there are search results, just
keep everything.

Change-Id: Ia5ef2b35b40b38fdcc95cf963d58bac46010e081