Add a link on static story page for StoryViewer on article page
Navigation between article with StoryViewer and static story page
is not easy. The various links on wiki link to the static story
page but there is no easy way to get to the real story viewer
or the related article from there.

this patch adds a simple link on top of the static story page
to the article with the story id in the url so that it opens
the story viewer.

Change-Id: I171b316cbb12f1431016f5ac3c667ac669e9ac66