Try to read the article id from the available page reference
By creating a new title object to get the article id,
it is possible that it read from a DB replica that doesn't
have know that page yet.

This code tries to use the title object already available
in the cpoParams but still tries another way if it's not
a Title instance.

Bug: T312106
Change-Id: I4ead96f47b855abbebd10de36837e44a3e509ada