Fix flow of select-text CTA
The select text CTA is using flex with align-items in order
to make sure it is in the center of the box. However, since the text
itself is composed of more than one node (one or more text node and
a span for the word Wikipedia with distinct styling), the nodes
are flex items that flow horizontally and wrap their own text.

The solution is to wrap the message in a span to ensure the CTA
always has a single child.

Also fix the message text and padding while at it.

Bug: T317338
Change-Id: I1b7997d3ea5a0f9d7ab921912cb2358d9d85a53d