Properly unwrap links
The previous solution for removing links was replacing them
with their text content. That works fine when they indeed
contain text but when they contain HTML markup it outputs
it as text, which led to visible HTML on the page.

Bonus: remove link to add article description from
article text selection view.

Bug: T352216
Change-Id: Icab86d917c44edc7dc72a6b8b8ff6d65731cbfcb